chore: combined eth_sign and personal_sign components
