+++ src/views/Home.vue
@@ -8,7 +8,6 @@ import Permissions from '@/components/endpoints/Permissions.vue'
 import Assets from '@/components/endpoints/Assets.vue'
 import Sign from '@/components/endpoints/Sign.vue'
 import SignTransaction from '@/components/endpoints/SignTransaction.vue'
-import PersonalSign from '@/components/endpoints/PersonalSign.vue'
 import CustomRelayer from '@/components/endpoints/CustomRelayer.vue'
 import Notifications from '@/components/Notification.vue'
 import ImportProfile from '@/components/endpoints/ImportProfile.vue'
@@ -47,10 +46,7 @@ const hasExtension = computed(() => !!window.lukso)
     </div>
     <div class="tile is-ancestor">
       <Sign />
-      <PersonalSign />
       <SignTransaction />
-    </div>
-    <div class="tile is-ancestor">
       <GetNetworkId />
     </div>
   </section>
